scanpy/pyproject.toml
[tool.poetry]
name = "scanpy-tutorials"
version = "0.1.0"
description = "Tutorials and experiments with Scanpy for single-cell analysis"
authors = ["Pantelis Vlachas <prvlachas@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
scanpy = "^1.10.2"
anndata = "^0.10.9"
numpy = "^1.24"
pandas = "^2.0"
matplotlib = "^3.8"
seaborn = "^0.13"
scikit-learn = "^1.4"
umap-learn = "^0.5"
h5py = "^3.9"
tqdm = "^4.66"
notebook = "^7.0"
jupyterlab = "^4.0"
ipywidgets = "^8.1"
pooch = "^1.8.2"
igraph = "^1.0.0"
scikit-image = "^0.25.2"
pyro-ppl = "^1.9.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
ruff = "^0.6"
pytest = "^8.0"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
fix = true

[tool.pytest.ini_options]
addopts = "--maxfail=1 --disable-warnings -q"
testpaths = ["tests"]
